Move GRS hashes to grs package
Add groestlcoin_hash to tox.ini
