+++ tox.ini
@@ -10,3 +10,4 @@ envlist = py27, py34, py35, py36, pypy
 commands = py.test tests
 deps =
     pytest
+    groestlcoin_hash
